Use jsx notation for selectors
